Move utility hooks into utils
