+++ packages/leva/src/components/Folder/Folder.tsx
@@ -4,7 +4,7 @@ import { StyledFolder, StyledWrapper, StyledContent } from './StyledFolder'
 import { isInput } from '../Leva/tree'
 import { join } from '../../utils'
 import { Control } from '../Control'
-import { useToggle } from '../../hooks'
+import { useToggle } from '../../utils/hooks'
 import { FolderSettings, Tree } from '../../types/'
 import { useStoreContext } from '../../context'
