+++ packages/leva/src/components/Leva/LevaRoot.tsx
@@ -4,7 +4,7 @@ import { StoreType } from '../../store'
 import { buildTree } from './tree'
 import { TreeWrapper } from '../Folder'
 
-import { useDeepMemo, useTransform, useVisiblePaths } from '../../hooks'
+import { useDeepMemo, useTransform, useVisiblePaths } from '../../utils/hooks'
 
 import { StyledRoot } from './StyledLeva'
 import { mergeTheme, globalStyles, LevaCustomTheme } from '../../styles'
