+++ packages/leva/src/components/Number/Number.tsx
@@ -3,7 +3,7 @@ import { NumberInput } from '../ValueInput'
 import { LevaInputProps } from '../../types/'
 import { InternalNumberSettings } from './number-plugin'
 import { Label, Row } from '../UI'
-import { useDragNumber } from '../../hooks'
+import { useDragNumber } from '../../utils/hooks'
 import { RangeGrid, InnerNumberLabel } from './StyledNumber'
 import { RangeSlider } from './RangeSlider'
 import { useInputContext } from '../../context'
