+++ packages/leva/src/components/Vector/Vector.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { useValue } from '../../hooks'
+import { useValue } from '../../utils/hooks'
 import { sanitizeValue } from '../../utils'
 import { Number } from '../Number'
 import { InternalNumberSettings } from '../Number/number-plugin'
