+++ packages/leva/src/plugins/index.ts
@@ -4,7 +4,7 @@ export { createPlugin } from '../plugin'
 export { Row, Label, Portal } from '../components/UI'
 export { ValueInput } from '../components/ValueInput'
 export { Vector, getVectorPlugin } from '../components/Vector'
-export { useDrag, useDragNumber, useCanvas2d } from '../hooks/'
+export { useDrag, useDragNumber, useCanvas2d } from '../utils/hooks'
 export { useInputContext } from '../context'
 // export styling utilities
 export { styled, useTh } from '../styles'
