+++ packages/leva/src/utils/hooks/useCanvas.ts
@@ -1,5 +1,5 @@
 import { useEffect, useRef } from 'react'
-import { debounce } from '../utils'
+import { debounce } from '../../utils'
 
 export function useCanvas2d(
   fn: Function
