+++ packages/leva/src/utils/hooks/useVisiblePaths.ts
@@ -1,4 +1,4 @@
-import { StoreType } from '../store'
+import { StoreType } from '../../store'
 import shallow from 'zustand/shallow'
 
 /**
